Move remaining files in server/ folder
